Phase-0 closure: batch executed end-to-end; Phase-2 sealed+tooled
s0 F5 route=FALLBACK_65PAIR (36.4 GPU-h projected > 12h rule); s2 inert
extraction complete (cosmetic status bug fixed); s3 full-seq curve+family
check: July DeepScaleR dPR not re-extraction-stable (sign flip, mapping
NOT LICENSED), SV-ratio profile flat 0.984-0.986 = uniform ~1.5% shrink,
no tail contraction. Dispositions in freeze §4b; ledger updated.

Phase 2: prereg SEALED + A1 (fresh-task manifest, corrected: dynamic id
start _117) + A2 adjunct SEALED; ph2_manifest/ph2_executor tooling with
16/16 pre-spend tests; checkpoint provenance table; pt13b/ph0 Phase-0
calibrations; codex handoff in .codex/out. Volume-side artifacts (s3
npz/json, STAR1 inert) + F5 fallback (prospective/unrun) = next pod.

+++ src/chain_gen.py
@@ -121,7 +121,7 @@ def load_model(
 
     model = AutoModelForCausalLM.from_pretrained(
         model_id,
-        dtype=torch_dtype,
+        torch_dtype=torch_dtype,  # 4.x name; 5.x accepts it (deprecated) — pod images vary
         device_map=device_map,
         trust_remote_code=True,
         quantization_config=quant_cfg,
